+++ interfacesd.tf
@@ -50,19 +50,21 @@ locals {
   }] : []
 
   interfaced_cloud_config_file_map = {
-    users = local.additional_users_map
+    users    = local.additional_users_map
     timezone = var.timezone
     write_files = flatten([
       local.additional_hosts_entries_cloud_init_write_files_map,
       local.additional_files_cloud_init_write_files_map,
       local.interfaced_network_config_file_map,
       local.interfaced_nameservers_file_map,
-      local.timezone_cloud_init_write_files_map
+      local.timezone_cloud_init_write_files_map,
+      local.packages_install_script_file_map
     ])
     runcmd = length(local.interfaced_nameservers_list) > 0 ? flatten([
       local.additional_hosts_entries_cloud_init_run_cmd_list,
-      var.additional_run_commands,
-      "systemctl enable resolvconf"
+      (var.upgrade_all_packages || length(var.additional_packages) > 0) && var.private_networks_only ? [".${local.packages_install_script_path}"] : [],
+      "systemctl enable resolvconf",
+      var.additional_run_commands
       ]) : flatten([
       local.additional_hosts_entries_cloud_init_run_cmd_list,
       var.additional_run_commands,