+++ outputs.tf
@@ -1,15 +1,3 @@
-# Common outputs
-
-output "result_file" {
-  value = local.result_user_data_file
-}
-
-output "result_hosts_file" {
-  value = local.additional_hosts_entries_file
-}
-
-# interfaces.d outputs
-
 /*
 Terraform module for creating Hetzner cloud compatible user-data file
 Copyright (C) 2021 Wojciech Szychta
@@ -27,6 +15,23 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
+
+# Common outputs
+
+output "result_file" {
+  value = local.result_user_data_file
+}
+
+output "result_hosts_file" {
+  value = local.additional_hosts_entries_file
+}
+
+output "packages_install_script" {
+  value = local.packages_install_script_file
+}
+
+# interfaces.d outputs
+
 output "result_interfacesd_file_map" {
   value = local.interfaced_cloud_config_file_map
 }
@@ -57,4 +62,22 @@ output "netplan_network_merge_script" {
 
 output "result_ifcfg_cloud_config_map" {
   value = local.ifcfg_cloud_config_file_map
+}
+
+output "ifcfg_network_config_files_map" {
+  value = local.ifcfg_network_config_files_map
+}
+
+output "ifcfg_network_routes_files_map" {
+  value = local.ifcfg_network_routes_files_map
+}
+
+# keyfile outputs
+
+output "result_keyfile_cloud_config_map" {
+  value = local.keyfile_cloud_config_file_map
+}
+
+output "keyfile_network_config_files_map" {
+  value = local.keyfile_network_config_files_map
 }